Introduce check_dimensions and an optional transform param
